update container state to DELETED

+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/fsck/ContainerHealthTask.java
@@ -22,7 +22,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
@@ -35,7 +34,6 @@ import org.apache.hadoop.hdds.scm.container.ContainerManager;
 import org.apache.hadoop.hdds.scm.container.ContainerNotFoundException;
 import org.apache.hadoop.hdds.scm.container.ContainerReplica;
 import org.apache.hadoop.hdds.scm.container.common.helpers.ContainerWithPipeline;
-import org.apache.hadoop.hdds.scm.ha.SCMHAUtils;
 import org.apache.hadoop.ozone.common.statemachine.InvalidStateTransitionException;
 import org.apache.hadoop.ozone.recon.persistence.ContainerHealthSchemaManager;
 import org.apache.hadoop.ozone.recon.scm.ReconScmTask;
@@ -83,15 +81,6 @@ public class ContainerHealthTask extends ReconScmTask {
     this.placementPolicy = placementPolicy;
     this.containerManager = containerManager;
     interval = reconTaskConfig.getMissingContainerTaskInterval().toMillis();
-  }
-
-  /**
-   * Container status in SCM enum.
-   */
-  public enum ContainerStatusInSCM {
-    DELETED,
-    NOT_DELETED,
-    NOT_FOUND
   }
 
   @Override
@@ -186,24 +175,24 @@ public class ContainerHealthTask extends ReconScmTask {
           if (ContainerHealthRecords
               .retainOrUpdateRecord(currentContainer, rec
               )) {
-            if (currentContainer.isMissing()) {
-              // Check if the missing container is deleted in SCM
-              ContainerStatusInSCM containerStatus =
-                  containerDeletedInSCM(currentContainer.getContainer());
-              if (!Objects.equals(containerStatus,
-                  ContainerStatusInSCM.NOT_DELETED)) {
-                rec.delete();
-
-                if (Objects.equals(containerStatus,
-                    ContainerStatusInSCM.NOT_FOUND)) {
-                  try {
-                    containerManager.deleteContainer(
-                        currentContainer.getContainer().containerID());
-                  } catch (IOException ex) {
-                    LOG.error("Unable to delete container during " +
-                              "periodic container health task.");
-                  }
+            // Check if the missing container is deleted in SCM
+            if (currentContainer.isMissing() &&
+                containerDeletedInSCM(currentContainer.getContainer())) {
+              rec.delete();
+              ContainerID id = currentContainer.getContainer().containerID();
+              try {
+                if (containerManager.getContainer(id)
+                        .getState() == HddsProtos.LifeCycleState.CLOSING) {
+                  containerManager.updateContainerState(id,
+                      HddsProtos.LifeCycleEvent.CLOSE);
                 }
+                containerManager.updateContainerState(id,
+                    HddsProtos.LifeCycleEvent.DELETE);
+                containerManager.updateContainerState(id,
+                    HddsProtos.LifeCycleEvent.CLEANUP);
+              } catch (InvalidStateTransitionException | IOException ex) {
+                LOG.error("Exception while updating " +
+                          "deleted container state: " + ex);
               }
             }
             existingRecords.add(rec.getContainerState());
@@ -237,9 +226,7 @@ public class ContainerHealthTask extends ReconScmTask {
         return;
       }
       // For containers deleted in SCM, we sync the container state here.
-      if (h.isMissing() &&
-          Objects.equals(containerDeletedInSCM(container),
-              ContainerStatusInSCM.DELETED)) {
+      if (h.isMissing() && containerDeletedInSCM(container)) {
         return;
       }
       containerHealthSchemaManager.insertUnhealthyContainerRecords(
@@ -250,8 +237,7 @@ public class ContainerHealthTask extends ReconScmTask {
     }
   }
 
-  private ContainerStatusInSCM containerDeletedInSCM(
-      ContainerInfo containerInfo) {
+  private boolean containerDeletedInSCM(ContainerInfo containerInfo) {
     try {
       ContainerWithPipeline containerWithPipeline =
           scmClient.getContainerWithPipeline(containerInfo.getContainerID());
@@ -268,22 +254,16 @@ public class ContainerHealthTask extends ReconScmTask {
           containerManager.updateContainerState(containerInfo.containerID(),
               HddsProtos.LifeCycleEvent.CLEANUP);
         }
-        return ContainerStatusInSCM.DELETED;
+        return true;
       }
     } catch (InvalidStateTransitionException e) {
       LOG.error("Failed to transition Container state while processing " +
           "container in Container Health task", e);
     } catch (IOException e) {
-      Throwable t = SCMHAUtils.unwrapException(e);
-      if (t instanceof ContainerNotFoundException) {
-        LOG.error("Container not present in SCM", t);
-        return ContainerStatusInSCM.NOT_FOUND;
-      } else {
-        LOG.error("Got exception while processing container in" +
-            " Container Health task", e);
-      }
+      LOG.error("Got exception while processing container in" +
+          " Container Health task", e);
     }
-    return ContainerStatusInSCM.NOT_DELETED;
+    return false;
   }
 
   /**
